backend/requirements.txt
python-dotenv
psycopg2-binary
Django>=4.2
djangorestframework
django-cors-headers
web3
cryptography